requirements.txt
aiohttp==3.8.1
amqp==2.5.2
appdirs==1.4.4
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
bcrypt==3.2.0
billiard==3.6.4.0
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
cfgv==3.1.0
chardet==3.0.4
cryptography==2.9
distlib==0.3.0
django-cors-headers==3.11.0
django-redis==4.11.0
filelock==3.0.12
identify==1.4.14
idna==2.9
importlib-metadata==1.6.0
kombu==4.6.8
multidict==4.7.5
nodeenv==1.3.5
paramiko==2.7.1
pre-commit==2.2.0
psycopg2==2.8.4
pycparser==2.20
pyjq==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
redis==3.4.1
requests==2.23.0
sentry-sdk==0.16.2
six==1.14.0
slackclient==2.5.0
sqlparse==0.3.1
text-unidecode==1.3
toml==0.10.0
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.16
yarl==1.4.2
zipp==3.1.0
djangorestframework==3.11.0
PyJWT==2.3.0
Django==4.0.1
djangorestframework-simplejwt==4.4.0
django-timezone-field==4.0
djangorestframework-camel-case==1.1.2
factory_boy==2.12.0
pycryptodome==3.9.7
frozendict==1.2
django-nested-admin==3.3.0
django-crispy-forms==1.9.0
django-cors-middleware==1.5.0
django-cacheback[celery]==2.0.0
Faker==4.1.0
codecov==2.1.4
packaging==20.4
Fabric3==1.14.post1
humanize==3.2.0
mixpanel==4.8.1
